+++ Products/DataGridField/Column.py
@@ -23,13 +23,13 @@ class Column(object):
 
     security = ClassSecurityInfo()
 
-    def __init__(self, label, col_description=None, default=None, default_method=None, label_msgid=None, required=False):
+    def __init__(self, label, col_description=None, default=None, default_method=None, required=False):
         """ Create a column
 
             @param label User visible name
             @param col_description General description for the column (when label is  not enough)
             @param default Default value for new rows
-            @param default_value Default function to generate the default value for new rows
+            @param default_method Default function to generate the default value for new rows
             @param required Set to True when values in this column are required
         """
         self.label = label
@@ -37,10 +37,6 @@ class Column(object):
         self.default = default
         self.default_method = default_method
         self.required = required
-
-        if label_msgid is None:
-            label_msgid = label
-            self.label_msgid = label_msgid
 
     security.declarePublic('getLabel')
     def getLabel(self, context, widget=None):