+++ Products/DataGridField/FixedColumn.py
@@ -23,17 +23,6 @@ class FixedColumn(Column):
     """
     security = ClassSecurityInfo()
 
-    def __init__(self, label, col_description=None, default=None,
-                 default_method=None, visible=True, required=False):
-        """ Create a column
-
-            @param visible Hide column from displaying by setting this to False
-        """
-        Column.__init__(self, label, col_description=col_description,
-                        default=default, default_method=default_method,
-                        required=required)
-        self.visible = visible
-
     security.declarePublic('getMacro')
     def getMacro(self):
         """ Return macro used to render this column in view/edit """